[BUGFIX] Dimensions are not filled for YouTube movies
For calculating widths and heights in a gallery, like in Text and Media, the dimensions of movies are necessary. YouTube does not provide this data in gdata, as used in the YouTubeHelper, but does provide dimensions in oembed.

This patch adds a call to oembed next to the existing gdata call, for the only purpose to get the dimensions. The width and height meta data for the YouTube movie is filled with these dimensions.